Created project-item component Created index.tsx for constants value. Move constants value definition(project,certification,faq,work-history,skill-set) to constants file.
