+++ components/ui/certification-item.tsx
@@ -1,7 +1,7 @@
 import Link from "next/link";
 import React from "react";
 
-interface CertificationItemProps {
+type CertificationItemProps ={
   id: number; // index of certification list
   name: string; // name of certification
   organization: string; // organization name
@@ -44,3 +44,4 @@ const CertificationItem = ({ id,name,organization="unknown", date="",verifyLink,
 };
 
 export default CertificationItem;
+export type { CertificationItemProps };
