+++ components/ui/project-item.tsx
@@ -1,0 +1,81 @@
+"use client"
+import Image from 'next/image'
+import Link from 'next/link'
+import { ExternalLink,Github} from 'lucide-react'
+import {projects} from '@/constants'
+import profilePic from "public/images/profile/developer-pic-1.png";
+
+const tmpPic = "/public/images/profile/developer-pic-1.png";
+
+type Project = {
+  id: string
+  title: string
+  description: string
+  image: string
+  technologies: string[]
+  githubUrl: string
+  liveUrl: string
+}
+
+const ProjectPage = () => {
+  return (
+    <div className="min-h-screen bg-gray-100 py-12 px-4 sm:px-6 lg:px-8">
+      <div className="max-w-7xl mx-auto">
+        <h1 className="text-4xl font-bold text-center text-gray-900 mb-12">My Projects</h1>
+        <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-8">
+          {projects.map((project) => (
+            <div key={project.id} className="bg-white rounded-lg shadow-lg overflow-hidden">
+              <Image
+                src={project.image}
+                alt={project.title}
+                width={600}
+                height={400}
+                className="w-full h-48 object-cover"
+              />
+              <div className="p-6">
+                <h2 className="text-2xl font-bold text-gray-900 mb-2">{project.title}</h2>
+                <p className="text-gray-600 mb-4">{project.description}</p>
+                <div className="mb-4">
+                  <h3 className="text-lg font-semibold text-gray-900 mb-2">Technologies:</h3>
+                  <div className="flex flex-wrap gap-2">
+                    {project.technologies.map((tech, index) => (
+                      <span
+                        key={index}
+                        className="bg-blue-100 text-blue-800 text-xs font-medium px-2.5 py-0.5 rounded"
+                      >
+                        {tech}
+                      </span>
+                    ))}
+                  </div>
+                </div>
+                <div className="flex justify-between items-center">
+                  <Link
+                    href={project.githubUrl}
+                    target="_blank"
+                    rel="noopener noreferrer"
+                    className="text-gray-600 hover:text-gray-900 transition-colors"
+                  >
+                    <Github className="w-6 h-6" />
+                    <span className="sr-only">GitHub Repository</span>
+                  </Link>
+                  <Link
+                    href={project.liveUrl}
+                    target="_blank"
+                    rel="noopener noreferrer"
+                    className="text-gray-600 hover:text-gray-900 transition-colors"
+                  >
+                    <ExternalLink className="w-6 h-6" />
+                    <span className="sr-only">Live Demo</span>
+                  </Link>
+                </div>
+              </div>
+            </div>
+          ))}
+        </div>
+      </div>
+    </div>
+  )
+}
+
+export default ProjectPage
+export type {Project}
